pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Calendar"
version = "0.0.1"
description = "Console calendar"
authors = [{ name = "Mikołaj Marsy", email = "mikolaj.marsy@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.7.0",
    "isort>=5.12.0",
    "black>=23.0.0",
    "pre-commit",
    "pytest",
    "pytest-mock",
    "pytest-cov",
    "pydoc-markdown"
]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true


[tool.setuptools.packages.find]
where = [
    "src"
]

# CLI entry point (if you want a command-line tool)
[project.scripts]
    "calendar_setup" = "scripts:setup"
    "ctor" = "scripts:ctor"
    "capp" = "scripts:cli"
